requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.4
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==4.0.6
django-allauth==0.51.0
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-cuser==2017.3.16
django-dotenv==1.4.2
django-import-export==2.8.0
django-modeltranslation==0.18.4
django-timezone-field==5.0
django-tinymce==3.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
et-xmlfile==1.1.0
idna==3.3
kombu==5.2.4
MarkupPy==1.14
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
pdfkit==1.0.0
Pillow==9.2.0
plotly==5.9.0
prompt-toolkit==3.0.30
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
tenacity==8.0.1
tzdata==2022.1
urllib3==1.26.10
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
